Gut out old Engine Logger code
