+++ YARG.Core/Engine/BaseEngine.Generic.cs
@@ -305,8 +305,6 @@ namespace YARG.Core.Engine
 
             EngineStats.Reset();
 
-            EventLogger.Clear();
-
             foreach (var note in Notes)
             {
                 note.ResetNoteState();
