+++ YARG.Core/Replays/Analyzer/AnalysisResult.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-using YARG.Core.Engine;
-using YARG.Core.Engine.Logging;
+﻿using YARG.Core.Engine;
 
 namespace YARG.Core.Replays.Analyzer
 {
@@ -11,7 +9,5 @@ namespace YARG.Core.Replays.Analyzer
         public BaseStats Stats;
 
         public int ScoreDifference;
-
-        public EngineEventLogger? EventLogger;
     }
 }
