+++ YARG.Core/Replays/ReplayFrame.cs
@@ -1,7 +1,6 @@
 using System;
 using YARG.Core.Engine;
 using YARG.Core.Engine.Guitar;
-using YARG.Core.Engine.Logging;
 using YARG.Core.Input;
 
 namespace YARG.Core.Replays
@@ -14,9 +13,6 @@ namespace YARG.Core.Replays
         public int                  InputCount;
         public GameInput[]          Inputs;
 
-        // Unused since version 6
-        public EngineEventLogger    EventLog;
-
         // Disabling this because it looks ugly with the initializers lol
         // ReSharper disable once ConvertConstructorToMemberInitializers
         public ReplayFrame()
@@ -24,7 +20,6 @@ namespace YARG.Core.Replays
             EngineParameters = new GuitarEngineParameters();
             Stats = new GuitarStats();
             Inputs = Array.Empty<GameInput>();
-            EventLog = new EngineEventLogger();
         }
     }
 }
