+++ YARG.Core/Replays/v012/V012ReplaySerializer.Sections.cs
@@ -3,7 +3,6 @@ using System.IO;
 using YARG.Core.Engine;
 using YARG.Core.Engine.Drums;
 using YARG.Core.Engine.Guitar;
-using YARG.Core.Engine.Logging;
 using YARG.Core.Engine.ProKeys;
 using YARG.Core.Engine.Vocals;
 using YARG.Core.Game;
